+++ psutil/arch/netbsd/mem.c
@@ -22,8 +22,15 @@
 #include "../../arch/all/init.h"
 
 
-// Virtual memory stats, taken from:
+// Virtual memory stats for NetBSD using VM_UVMEXP2 and VM_METER.
 // https://github.com/zabbix/zabbix/blob/master/src/libs/zbxsysinfo/netbsd/memory.c
+//
+// Sources:
+//   cached  = (filepages + execpages + anonpages) << pageshift  [btop]
+//   buffers = 0 [follow OpenBSD's psutil implementation]
+//   shared  = (t_vmshr + t_rmshr) * pagesize [vmtotal]
+//   used    = (active + wired) << pageshift [top/btop]
+//   avail   = total - used [htop/btop]
 PyObject *
 psutil_virtual_mem(PyObject *self, PyObject *args) {
     struct uvmexp_sysctl uv;
@@ -34,8 +41,6 @@ psutil_virtual_mem(PyObject *self, PyObject *args) {
     unsigned long long buffers, shared, used, avail;
     double percent;
     long pagesize = psutil_getpagesize();
-    FILE *f;
-    char line[256];
     PyObject *dict = PyDict_New();
 
     if (dict == NULL)
@@ -46,34 +51,21 @@ psutil_virtual_mem(PyObject *self, PyObject *args) {
     if (psutil_sysctl(vmmeter_mib, 2, &vmdata, sizeof(vmdata)) != 0)
         goto error;
 
-    // get buffers from /proc/meminfo
-    buffers = 0;
-    f = fopen("/proc/meminfo", "r");
-    if (f != NULL) {
-        while (fgets(line, sizeof(line), f) != NULL) {
-            if (strncmp(line, "Buffers:", 8) == 0) {
-                sscanf(line + 8, "%llu", &buffers);
-                break;
-            }
-        }
-        fclose(f);
-        buffers *= 1024;
-    }
-
+    // https://github.com/aristocratos/btop/blob/v1.4.6/src/netbsd/btop_collect.cpp#L720
     total = (unsigned long long)uv.npages << uv.pageshift;
     free = (unsigned long long)uv.free << uv.pageshift;
     active = (unsigned long long)uv.active << uv.pageshift;
     inactive = (unsigned long long)uv.inactive << uv.pageshift;
     wired = (unsigned long long)uv.wired << uv.pageshift;
-    // Also in /proc/meminfo as MemShared
-    shared = (unsigned long long)(vmdata.t_vmshr + vmdata.t_rmshr) * pagesize;
-
     // Note: zabbix does not include anonpages, but that doesn't match
     // the "Cached" value in /proc/meminfo.
     // https://github.com/zabbix/zabbix/blob/af5e0f8/src/libs/
     //   zbxsysinfo/netbsd/memory.c#L182
     cached = (unsigned long long)(uv.filepages + uv.execpages + uv.anonpages)
              << uv.pageshift;
+    shared = (unsigned long long)(vmdata.t_vmshr + vmdata.t_rmshr) * pagesize;
+    // XXX: Still being determined.
+    buffers = (unsigned long long)uv.filepages << uv.pageshift;
 
     // Before avail was calculated as (inactive + cached + free), same
     // as zabbix, but it turned out it could exceed total (see #2233),